+++ CLAUDE.md
@@ -22,7 +22,7 @@
     `#pdf-exporter-panel` div plus a `<link>` to `css/starter.css` whose `onload` fires the import) and
     `assets/side-panel.js` mounts React into a shadow root of it. It reads its data from the same internal
     REST endpoints the export dialog uses; the Java side substitutes nothing but the bundle version.
-    Its CSS is `ui/src/formext/side-panel.css`.
+    Its CSS is `ui/src/sidepanel/side-panel.css`.
   - **"Export to PDF" dialog** - `assets/export-popup.js` exporting `openExportPopup({documentType})`,
     imported on click by `js/starter.js` (document editor toolbar), `js/live-reports.js` (report toolbar)
     and `ExportToPdfButtonRenderer` (the report widget button). It appends its own host to the page body and
