+++ src/main/resources/webapp/pdf-exporter/html/sidePanelContent.html
@@ -1,6 +1,6 @@
 <!-- Document Properties side panel. Polarion (PdfExporterFormExtension) injects this fragment into the
   document editor's Document Properties pane. On load the React side-panel bundle is imported and mounts
-  into a SHADOW ROOT on the div below (see ui/src/formext/mount.tsx). It injects its own styles into that
+  into a SHADOW ROOT on the div below (see ui/src/sidepanel/mount.tsx). It injects its own styles into that
   shadow root - react-sbb-polarion's bundled stylesheet plus side-panel.css - so the panel's CSS is fully
   encapsulated and cannot clash with other extensions' panels on this shared page. The generic control CSS
   is intentionally NOT linked globally here any more (it lives in the shadow).
